Added a delete function to clear the qmckl_context instance.
Perhaps we want to organize garbage collection differently !? I thought it's a nice first commit to add a basic delete function to motivate the ways in which instances can be deleted. Obviously this is only a proposal to learn from. Note that no information is being sent (other than the fact that free has been called) about the contents of the _context_ instance after the call to the function.